+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-# 2.0.0-beta.0
+# 2.0.0-beta.1
+
+- Fix refresh config bug
+- Modify props (See README.md)